Refactor web*.py to server/*.py
