+++ sickbeard/helpers.py
@@ -1004,7 +1004,7 @@ By Pedro Jose Pereira Vieito <pvieito@gmail.com> (@pvieito)
 
 To add a new encryption_version:
   1) Code your new encryption_version
-  2) Update the last encryption_version available in webserve.py
+  2) Update the last encryption_version available in web.py
   3) Remember to maintain old encryption versions and key generators for retrocompatibility
 """
